Multiplatform-demo example
1. Make MPP example as a main example for Compose Multiplatform:
<image>
2. For iOS/Android/Web it will run as a list of examples, for Desktop we create separate configurations to speedup opening (for other targets we will figure something in the future probably)
