+++ compose/mpp/demo/build.gradle.kts
@@ -219,7 +219,8 @@ if (System.getProperty("os.name") == "Mac OS X") {
 
 tasks.create("runDesktop", JavaExec::class.java) {
     dependsOn(":compose:desktop:desktop:jar")
-    main = "androidx.compose.mpp.demo.Main_desktopKt"
+    mainClass.set("androidx.compose.mpp.demo.Main_desktopKt")
+    args = listOfNotNull(project.findProperty("args")?.toString())
     systemProperty("skiko.fps.enabled", "true")
     val compilation = kotlin.jvm("desktop").compilations["main"]
     classpath =
