+++ compose/mpp/demo/src/commonMain/kotlin/androidx/compose/mpp/demo/App.kt
@@ -1,138 +1,77 @@
 package androidx.compose.mpp.demo
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.material.Button
-import androidx.compose.material.ButtonDefaults
-import androidx.compose.material.Checkbox
-import androidx.compose.material.RadioButton
-import androidx.compose.material.Switch
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.Icon
+import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
-import androidx.compose.material.TextField
-import androidx.compose.runtime.Applier
+import androidx.compose.material.TopAppBar
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.Stable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.runtime.withFrameNanos
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.platform.LocalClipboardManager
-import androidx.compose.ui.platform.LocalUriHandler
-import kotlinx.serialization.Serializable
 
-@Serializable // Not used at the moment, but added here to test compose with kotlinx.serialization compilation
-public class Id(public val id: Applier<String>)
+class App(
+    private val initialScreenName: String? = null
+) {
+    private val MainScreen = Screen("Demo") { Main() }
 
-@Composable // Not used at the moment, but added here to test compose with kotlinx.serialization compilation
-fun Abc(id: Id) {
-    println("Id = $id")
-}
+    private val screens = listOf(
+        Screen("Example1") { Example1() },
+    )
 
-@Composable
-fun myContent() {
-    var tick by remember { mutableStateOf(false) }
-    var selected by remember { mutableStateOf(false) }
-    var clutz by remember { mutableStateOf(false) }
-    var switched by remember { mutableStateOf(false) }
-    var textFromClipboard by remember { mutableStateOf("click for clipboard") }
-    var textFieldState by remember { mutableStateOf("I am TextField") }
-    val clipboard = LocalClipboardManager.current
-    val uriHandler = LocalUriHandler.current
-    Column {
-        TextField(
-            value = textFieldState,
-            onValueChange = {
-                textFieldState = it
-            }
-        )
-        Box(
-            modifier = Modifier
-                .padding(16.dp)
-                .background(color = if (selected) Color.Gray else Color.Red)
-                .width(100.dp).height(100.dp)
-                .clickable {
-                    println("Red box: clicked")
-                }
-        ) {
-            Box(
-                modifier = Modifier
-                    .padding(16.dp)
-                    .background(color = if (tick) Color.Green else Color.Blue)
-                    .width(20.dp).height(20.dp)
-                    .clickable {
-                        println("Small box: clicked")
+    private class Screen(val title: String, val content: @Composable () -> Unit)
+
+    private var screen: Screen by mutableStateOf(
+        if (initialScreenName != null) {
+            screens.find { it.title == initialScreenName }!!
+        } else {
+            MainScreen
+        }
+    )
+
+    @Composable
+    fun Content() {
+        Scaffold(
+            topBar = {
+                TopAppBar(
+                    title = { Text(screen.title) },
+                    navigationIcon = if (screen != MainScreen) {
+                        {
+                            Icon(
+                                Icons.Filled.ArrowBack,
+                                contentDescription = "Back",
+                                modifier = Modifier.clickable { screen = MainScreen }
+                            )
+                        }
+                    } else {
+                        null
                     }
-            )
-        }
-        Spacer(
-            Modifier.width(200.dp)
-                .height(if (clutz) 4.dp else 12.dp)
-                .background(color = if (clutz) Color.DarkGray else Color.Magenta)
-        )
-        Button(
-            modifier = Modifier
-                .padding(16.dp),
-            onClick = {
-                println("Button clicked!")
-                tick = !tick
+                )
             }
         ) {
-            Text(if (switched) "🦑 press 🐙" else "Press me!")
+            screen.content()
         }
-        Row {
-            RadioButton(
-                modifier = Modifier
-                    .padding(16.dp),
-                selected = selected,
-                onClick = {
-                    println("RadioButton clicked!")
-                    selected = !selected
-                }
-            )
+    }
 
-            Checkbox(
-                checked = clutz,
-                modifier = Modifier.padding(16.dp),
-                onCheckedChange = { clutz = !clutz }
-            )
-        }
-        Switch(
-            modifier = Modifier
-                .padding(16.dp),
-            checked = switched,
-            onCheckedChange = { switched = it }
-        )
-        Row {
-            Button(
-                modifier = Modifier.padding(16.dp),
-                onClick = {
-                    uriHandler.openUri("https://kotlinlang.org")
-                },
-            ) {
-                Text("Open URL")
-            }
-            Button(
-                modifier = Modifier.padding(16.dp),
-                onClick = {
-                    textFromClipboard = clipboard.getText()?.text ?: "clipboard is empty"
-                },
-                colors = ButtonDefaults.buttonColors(backgroundColor = Color.Gray)
-            ) {
-                Text(textFromClipboard)
+    @Composable
+    fun Main() {
+        LazyColumn(Modifier.fillMaxSize()) {
+            items(screens) {
+                Text(it.title, Modifier.clickable { screen = it }.padding(16.dp).fillMaxWidth())
             }
         }
-
     }
 }
